Ported more stuff from Slick2D, tweaked CoordinateTranslator to work with LibGdx coordinate system. Upped enemy speed to compensate for fixed frame rate compared to Slick's un fixed frame rate.

+++ core/src/Entity/Entity.java
@@ -43,7 +43,7 @@ public class Entity
     }
     
     /* Blank update method for inheriting class to implement*/
-    public void update(int t)
+    public void update(float t)
     {
         
     }
